We don't have to have the Python 2 version of diff_match_patch any longer
